Extract single-query request controls
